Feat/play

+++ src/components/PlayerCard/PlayerCard.tsx
@@ -1,5 +1,4 @@
-import { Player } from "@/types/player";
-import { HeartFilledIcon } from "@radix-ui/react-icons";
+import { HeartFilledIcon, TargetIcon } from "@radix-ui/react-icons";
 import { Badge } from "../ui/badge";
 import { Avatar, AvatarFallback, AvatarImage } from "../ui/avatar";
 import { motion } from "framer-motion";
@@ -22,12 +21,14 @@ import IconAvatar from "../IconAvatar/IconAvatar";
 import { useTranslation } from "react-i18next";
 import { DBEffect, Effect } from "@/types/effect";
 import EffectCard from "../EffectCard/EffectCard";
+import { Player } from "@/types/player";
 
 type Props = {
   player: Player;
   expanded: boolean;
-  onRemove: (playerId: Player["id"]) => void;
+  onRemove?: (playerId: Player["id"]) => void;
   onEdit: (player: Player) => void;
+  onSelectToken?: (playerId: Player["id"]) => void;
   onRemoveImmunity: (
     playerId: Player["id"],
     immunityId: DBImmunity["id"],
@@ -50,6 +51,7 @@ function PlayerCard({
   expanded,
   onEdit,
   onRemove,
+  onSelectToken,
   onEditEffect,
   onEditImmunity,
   onEditResistance,
@@ -70,7 +72,7 @@ function PlayerCard({
   );
 
   function handleRemovePlayer() {
-    onRemove(player.id);
+    if (onRemove) onRemove(player.id);
   }
 
   function handleEditPlayer() {
@@ -93,6 +95,17 @@ function PlayerCard({
     <DropdownMenuContent className="w-56">
       <DropdownMenuLabel>{player.name}</DropdownMenuLabel>
       <DropdownMenuSeparator />
+      {onSelectToken && (
+        <>
+          <DropdownMenuItem onClick={() => onSelectToken(player.id)}>
+            <div className="flex items-center gap-2">
+              <TargetIcon />
+              <span>{t("selectToken")}</span>
+            </div>
+          </DropdownMenuItem>
+          <DropdownMenuSeparator />
+        </>
+      )}
       <DropdownMenuGroup>
         <DropdownMenuItem onClick={onOpenEffectsCatalog}>
           {t("addEffect")}
@@ -184,7 +197,11 @@ function PlayerCard({
                     <ImmunityCard
                       key={`player-${player.id}-immunity-${immunity.id}`}
                       immunity={immunity}
-                      onRemove={() => handleRemoveImmunity(immunity.id)}
+                      onRemove={
+                        onRemove
+                          ? () => handleRemoveImmunity(immunity.id)
+                          : undefined
+                      }
                       onEdit={onEditImmunity}
                     />
                   ))}
@@ -228,7 +245,7 @@ function PlayerCard({
                 }
               >
                 <div className="flex w-full flex-col gap-4">
-                  {positiveEffects.map((effect) => (
+                  {positiveEffects.map((effect: Effect) => (
                     <EffectCard
                       key={`player-${player.id}-effect-${effect.id}`}
                       effect={effect}
@@ -252,7 +269,7 @@ function PlayerCard({
                 }
               >
                 <div className="flex w-full flex-col gap-4">
-                  {negativeEffects.map((effect) => (
+                  {negativeEffects.map((effect: Effect) => (
                     <EffectCard
                       key={`player-${player.id}-effect-${effect.id}`}
                       effect={effect}
@@ -283,34 +300,40 @@ function PlayerCard({
                 </span>
               </div>
 
-              <div className="flex items-center gap-[1px] rounded-md bg-gray-200 px-1 py-0.5 text-sm font-bold text-black">
+              <div className="flex w-full items-center justify-center gap-[1px] rounded-md bg-gray-200 px-1 py-0.5 text-sm font-bold text-black">
                 <span>{player.health}</span>
                 <span>/</span>
                 <span>{player.max_health}</span>
               </div>
 
-              <div className="flex w-full justify-between gap-2">
-                {player.effects.filter((effect) => effect.type === "positive")
-                  .length > 0 && (
-                  <div className="w-full rounded-md bg-emerald-500 px-1 py-0.5 text-sm font-bold text-white">
-                    {
-                      player.effects.filter(
-                        (effect) => effect.type === "positive",
-                      ).length
-                    }
+              {player.effects.length > 0 && (
+                <>
+                  <div className="flex w-full justify-between gap-2">
+                    {player.effects.filter(
+                      (effect: Effect) => effect.type === "positive",
+                    ).length > 0 && (
+                      <div className="w-full rounded-md bg-emerald-500 px-1 py-0.5 text-sm font-bold text-white">
+                        {
+                          player.effects.filter(
+                            (effect: Effect) => effect.type === "positive",
+                          ).length
+                        }
+                      </div>
+                    )}
+                    {player.effects.filter(
+                      (effect: Effect) => effect.type === "negative",
+                    ).length > 0 && (
+                      <div className="w-full rounded-md bg-red-500 px-1 py-0.5 text-sm font-bold text-white">
+                        {
+                          player.effects.filter(
+                            (effect: Effect) => effect.type === "negative",
+                          ).length
+                        }
+                      </div>
+                    )}
                   </div>
-                )}
-                {player.effects.filter((effect) => effect.type === "negative")
-                  .length > 0 && (
-                  <div className="w-full rounded-md bg-red-500 px-1 py-0.5 text-sm font-bold text-white">
-                    {
-                      player.effects.filter(
-                        (effect) => effect.type === "negative",
-                      ).length
-                    }
-                  </div>
-                )}
-              </div>
+                </>
+              )}
             </button>
           </DropdownMenuTrigger>
